handle moving and scrolling separately
only lf/ri should be scrolling, and only if they start on the border of
the scroll region and try to move off of it. also, all absolute movement
should be absolute in the context of the full terminal, but relative
vertical movement should be clamped to the scroll regions.