allow the cursor to go off the end of the line
vt100_screen_move_to clamps the cursor to being within the screen, but
when drawing text to the screen, the cursor is allowed to move to one
past the last column, and not wrap until the next character is written
(in this case, the cursor is still drawn in the last column, but the
internal terminal state is different). updating the row and column
manually allows us to position the column outside of the normal bounds
(which we otherwise don't want to allow vt100_screen_move_to to do).